+++ src/main/kotlin/dev/frozenmilk/dairy/calcified/hardware/encoder/Encoders.kt
@@ -12,7 +12,7 @@ class Encoders internal constructor(module: CalcifiedModule) : CalcifiedDeviceMa
 	/**
 	 * if the port is empty, makes a new [TicksEncoder], else, overrides the encoder on the port
 	 */
-	fun getTicksEncoder(port: Byte): TicksEncoder {
+	fun getTicksEncoder(port: Int): TicksEncoder {
 		// this is pretty much the same as the motors, as the encoders match the motors
 		// checks to confirm that the encoder port is validly in range
 		if (port !in LynxConstants.INITIAL_MOTOR_PORT until LynxConstants.INITIAL_MOTOR_PORT + LynxConstants.NUMBER_OF_MOTORS) throw IllegalArgumentException("$port is not in the acceptable port range [${LynxConstants.INITIAL_MOTOR_PORT}, ${LynxConstants.INITIAL_MOTOR_PORT + LynxConstants.NUMBER_OF_MOTORS - 1}]")
@@ -27,20 +27,20 @@ class Encoders internal constructor(module: CalcifiedModule) : CalcifiedDeviceMa
 	 *
 	 * @return Overrides the encoder on the port with a [Encoder] of the supplied type
 	 */
-	inline fun <reified T : Encoder<*>> getEncoder(lazySupplier: (TicksEncoder) -> T, port: Byte): T {
+	inline fun <reified T : Encoder<*>> getEncoder(lazySupplier: (TicksEncoder) -> T, port: Int): T {
 		val ticksEncoder = getTicksEncoder(port)
 		this[port] = lazySupplier(ticksEncoder)
 		return this[port] as? T ?: throw IllegalStateException("something went wrong while creating a new encoder, this shouldn't be reachable")
 	}
 
-	inline fun <reified T : Encoder<*>> getEncoder(port: Byte): T? {
+	inline fun <reified T : Encoder<*>> getEncoder(port: Int): T? {
 		return this[port] as? T
 	}
 
 	/**
 	 * overrides the encoder on the port with an [AngleEncoder], with the [ticksPerRevolution] and [wrapping] specified
 	 */
-	fun getAngleEncoder(port: Byte, wrapping: Wrapping, ticksPerRevolution: Double): AngleEncoder {
+	fun getAngleEncoder(port: Int, wrapping: Wrapping, ticksPerRevolution: Double): AngleEncoder {
 		return getEncoder({
 			AngleEncoder(it, wrapping, ticksPerRevolution)
 		}, port)
@@ -49,7 +49,7 @@ class Encoders internal constructor(module: CalcifiedModule) : CalcifiedDeviceMa
 	/**
 	 * overrides the encoder on the port with a [DistanceEncoder], with the [ticksPerUnit] specified
 	 */
-	fun getDistanceEncoder(port: Byte, unit: DistanceUnit, ticksPerUnit: Double): DistanceEncoder {
+	fun getDistanceEncoder(port: Int, unit: DistanceUnit, ticksPerUnit: Double): DistanceEncoder {
 		return getEncoder({
 			DistanceEncoder(it, unit, ticksPerUnit)
 		}, port)
